Introduced a new `ShortStock` project to handle short-selling
operations. Added new models (`MainInOutDetailResponse`,
`StockLimitPrice`, etc.) and repositories (`CandidateForShortRepository`,
`StockMainPowerRepository`) to support short-selling logic.

Enhanced `StockSelectorService` with methods for filtering
intraday-eligible stocks and tracking main power data. Updated
`FutureTraderService` with volume-fetching methods.

Integrated `Serilog` for logging and added `appsettings.json`
for configuration management. Updated `Core.csproj` and
`TwStockAutoTrade.sln` to include new files and the `ShortStock`
project. Added `YuantaCAPIDLL.dll` for API integration.

+++ Core/Model/YahooTechData.cs
@@ -23,7 +23,7 @@ namespace Core.Model
     }
     public class ChartModel
     {
-        public List<int> Timestamp { get; set; }
+        public List<long> Timestamp { get; set; }
         public IndicatorModel Indicators { get; set; }
     }
     public class IndicatorModel
@@ -36,6 +36,6 @@ namespace Core.Model
         public List<decimal> High { get; set; }
         public List<decimal> Low { get; set; }
         public List<decimal> Open { get; set; }
-        public List<int> Volume { get; set; }
+        public List<decimal> Volume { get; set; }
     }
 }
